Updated ZapRedraw document to work under 1.02.
Although the document has been updated to work under 1.02, this
does not mean that it's been converted completely. There are a lot
of structures that were added between 1.00 and 1.02 which can be
used to better effect. And there are some elements that are wrong
in the usage they have been put to (like sysvar to mean a general
variable, and lists where just bare paragraphs might do).

Outstanding issues section added to make it clearer what still
needs to happen.