Updated build to manage installation better outside of docker.
When not run as root, the jobs would fail because they could not
install anything. This has been made better by offering a prompt
to install, but only when there is a need to do so. So if you
already had the prerequisites installed it wouldn't try to install
anything through sudo.

The download and execution of the PrinceXML tool is now only
performed if the invoker has explicitly set an environment variable
to say that they're licensed. For non-commercial use licenses that
lets people set the variable for their own machines and on
automation if the output from the automation is not charged for.