Update the build with latest version and an index ready for GitHub pages.
The build system has been updated with features from the main PRM-in-XML
repository and the examples repository. This means that we now have an
index available which should be able to show you each of the document
styles, and we should be able to publish to the GitHub pages shortly.